requirements.txt
streamlit==1.28.0
pandas==2.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
plotly==5.17.0
